+++ src/components/SubjectCardList/SubjectCardListItem.tsx
@@ -9,7 +9,7 @@ import Flex from "../Flex";
 import OakLink from "../OakLink";
 import Card, { CardProps } from "../Card";
 import SubjectIcon from "../SubjectIcon";
-import { ProgrammesBySubject } from "../../pages/beta/teachers/key-stages/[keyStageSlug]/subjects";
+import { ProgrammesBySubject } from "../../pages/beta/[viewType]/key-stages/[keyStageSlug]/subjects";
 import useAnalyticsPageProps from "../../hooks/useAnalyticsPageProps";
 
 export type SubjectCardListItemProps = Omit<CardProps, "children"> & {
@@ -88,8 +88,7 @@ const SubjectCardListItem: FC<SubjectCardListItemProps> = ({
                   {...primaryTargetProps}
                   page="unit-index"
                   viewType="teachers"
-                  programme={programmeSlug}
-                  //TODO add tracking
+                  programmeSlug={programmeSlug}
                 >
                   {subjectTitle}
                 </OakLink>
@@ -98,9 +97,8 @@ const SubjectCardListItem: FC<SubjectCardListItemProps> = ({
                   {...primaryTargetProps}
                   page="programme-index"
                   viewType="teachers"
-                  keyStage={keyStageSlug}
-                  subject={subjectSlug}
-                  //TODO: replace 'key stage 4' with variable from above
+                  keyStageSlug={keyStageSlug}
+                  subjectSlug={subjectSlug}
                   onClick={() => {
                     track.subjectSelected({
                       keyStageTitle: keyStageTitle as KeyStageTitleValueType,